+++ CHANGELOG.md
@@ -90,6 +90,8 @@ Group changes to describe their impact on the project, as follows:
 - **Proxy/MediaRelay:** parameters `module::MediaRelay/sdp-port-range-min` and `module::MediaRelay/sdp-port-range-max` are 
   merged into unique parameter `module::MediaRelay/sdp-port-range`. They still work but are deprecated, please use the new 
   parameter.
+- **Configuration:** Parameter 'tls-certificates-dir' in global config and transports URI will be removed in 2.6.
+  Use 'tls-certificates-file', 'tls-certificates-private-key' and 'tls-certificates-ca-file' instead.
 
 ### [Fixed]
 - **Proxy:**